navigator: restructure client side compnents
