+++ client/.env.local_dev
@@ -1,3 +1,3 @@
 
-VITE_API_URL=http://localhost:5000/api
+VITE_API_URL=http://0.0.0.0:5001/api
 VITE_NODE_ENV=local
