+++ client/src/components/Content/PrivacyPolicy.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+
+
+const PrivacyPolicy: React.FC = () => (
+    <div className="policy-page">
+      <div className="container">
+        <h1>Privacy Policy</h1>
+        <div className="policy-content">
+          <section>
+            <h2>Information We Collect</h2>
+            <p>We collect information you provide directly to us, such as when you create an account, complete worksheets, or contact us for support.</p>
+          </section>
+  
+          <section>
+            <h2>How We Use Your Information</h2>
+            <p>We use the information we collect to provide, maintain, and improve our services, process transactions, and communicate with you.</p>
+          </section>
+  
+          <section>
+            <h2>Information Sharing</h2>
+            <p>We do not sell, trade, or rent your personal information to third parties. We may share your information only in specific circumstances outlined in this policy.</p>
+          </section>
+  
+          <section>
+            <h2>Data Security</h2>
+            <p>We implement appropriate security measures to protect your personal information against unauthorized access, alteration, disclosure, or destruction.</p>
+          </section>
+  
+          <section>
+            <h2>Contact Us</h2>
+            <p>If you have questions about this Privacy Policy, please contact us at privacy@fairchancenavigator.com</p>
+          </section>
+        </div>
+      </div>
+    </div>
+  );
+
+  export default PrivacyPolicy;
