+++ client/src/components/Content/TermsConditions.tsx
@@ -1,0 +1,38 @@
+import React from "react";
+
+
+const TermsConditions: React.FC = () => (
+    <div className="policy-page">
+      <div className="container">
+        <h1>Terms and Conditions</h1>
+        <div className="policy-content">
+          <section>
+            <h2>Acceptance of Terms</h2>
+            <p>By accessing and using Fair Chance Navigator 2.0, you accept and agree to be bound by the terms and provision of this agreement.</p>
+          </section>
+  
+          <section>
+            <h2>Use License</h2>
+            <p>Permission is granted to temporarily access the materials on Fair Chance Navigator 2.0 for personal, non-commercial transitory viewing only.</p>
+          </section>
+  
+          <section>
+            <h2>User Accounts</h2>
+            <p>You are responsible for maintaining the confidentiality of your account and password and for restricting access to your computer.</p>
+          </section>
+  
+          <section>
+            <h2>Prohibited Uses</h2>
+            <p>You may not use our service for any unlawful purpose or to solicit others to perform prohibited conduct.</p>
+          </section>
+  
+          <section>
+            <h2>Limitation of Liability</h2>
+            <p>In no event shall Fair Chance Navigator 2.0 or its suppliers be liable for any damages arising out of the use or inability to use the materials.</p>
+          </section>
+        </div>
+      </div>
+    </div>
+  );
+
+  export default TermsConditions;
