+++ client/tsconfig.node.json
@@ -1,9 +1,9 @@
 {
   "compilerOptions": {
+    "jsx": "react",
     "composite": true,
     "module": "ESNext",
     "moduleResolution": "Node",
     "allowSyntheticDefaultImports": true
-  },
-  "include": ["vite.config.ts"]
+  }
 }
