+++ client/vite.config.js
@@ -9,7 +9,7 @@ export default defineConfig({
     port: 5173,
     proxy: {
       '/api': {
-        target: 'http://localhost:5000',
+        target: 'http://0.0.0.0:5001',
         changeOrigin: true,
         secure: false,
       },
